Add files via upload
The following stuff:

The shell script - it takes two arguments and the usage is as following:
bash HW1_bianca.sh <path to file with genes only> <path to all genes table>
- The initial files
- The output files:

the _additionalinfo.txt files lists the additional information from the big table
the redundancies.txt files list the redundant genes and their count from the two up and down only tables. I concatenated the file, then sorted it, then used uniq with -cd which lists duplicates and counts them and wrote the output to this new text file.